feat: add minimal implementation of a query command
This introduces a datalad redcap-query command which queries for
available instruments (data entry forms). The command could be
extended in the future with other query options, e.g. (general)
project info. Unfortunately I didn't see a way to query for available
reports (report IDs) in the API playground.

This implementation subclasses PyCap.Instruments to query for
instruments names and labels (available in API, not exposed in
PyCap). It also adds a custom results renderer to print the results as
a formatted table.

Introduces a dependency on prettytable.