feat: add credentials support to export_form
This adds credentials retrieval/updating to export_form. Credential
name can be given as a parameter. Related to #2.

Uses CredentialManager from DataLad-Next, and the way it is used is
based on create_sibling_webdav, also from DataLad-Next.

This adds a dependency on DataLad-next (currently the main branch, as
the CredentialManager.obtain() method is not yet part of any release),
and bumps DataLad version dependency to 0.18.1 (matching Next's)