refactor: switch to using DataLad Next's constraints
This changes the constraint declaration for parameters to the method
introduced in DataLad Next.

The most visible change is that we introduce EnsureURL, in which we
require that that the API URL has scheme (e.g. `https`),
netloc (e.g. `www.example.com`) and path (e.g. `/api/`). This seems
reasonable, considering that PyCap complains even if the url ends with
`/api` and not `/api/`.

One thing that will probably change soon is that currently we need to
act differently if the dataset parameter is None, see
https://github.com/datalad/datalad-next/issues/225

Closes #5