test: add test for check_ok_to_edit
This will check whether "ok to edit" is reported correctly for files
inside a dataset (ok), outside, or under a subdataset (not ok), and
for modified files (not ok).

Does not check the "needs unlocking" part of the tested function - I
am currently not sure how to account for it on crippled filesystems.