build: add readthedocs config requiring python 3.9
This should resolve an issue with Read the Docs builds where Python
API and CLI pages were not built - inspection of RtD build logs showed
that build_manpage & autosummary had ImportErrors converted to
warnings. RtD build used Python 3.7.

Using Python 3.9 for building docs seems to be a sensible choice, if
only because that's the current version in debian stable (although
minimal requirement for the extension is 3.8).

Our docs build relies on having our package installed. This apparently
happened by default with no RtD config, but now this has to be made
explicit in the config file.