build: retire requirements-devel.txt in favour of setup.cfg
The requirements file is redundant since we use setup.cfg.
This should allow local development installation with:
`pip install -e .[devel]`.

The requirements file was used in the github actions (docbuild and
crippledfs), in the latter also to separate the installation of
dependencies and package into separate stages. The action
specification is updated to use the updated method, without
requirements file.

The requirements could be split into devel & docs, leading to:
`pip install -e .[devel,docs]`, but for now I don't see the need.