test: switch from decorators to fixtures
This removes the use of with_tempfile and with_credential decorators
and introduces pytest fixtures instead. Two custom fixtures allow
tests to run with a credential manager that has a dummy token matching
a dummy realm, so tested commands can run with the implicit credential
lookup.

Because of fixtures, the use of unittest.mock.patch was changed from
decorator to context manager -- the decorator wants to supply the
first positional argument, so depending on argument order either the
decorator replaces the fixture or pytest complains about unknown
fixture. Or at least I couldn't make it work together. But in the end,
context manager also looks good.