refactor: draft a context manager for credential manager
This is a draft - if the approach is adopted, the context manager
would move to utils.

The context manager is implemented using contextlib.contextmanager
utility decorator, not a class. It handles credential retrieval
and (if the with block succeeds) also credential update. It reraises
the two errors we can get (AssertionError or RedcapError), maintaining
previous behaviour.

However, if the errors were to be converted to DataLad result records,
it would be quite clumsy. My attempts to yield the result within the
generator caused error: generator didn't stop after throw(). Doing it
properly means an awkward try - with -- except.