Make sure that sphinx RTD theme gets installed in RTD build
This adds options.extras_require.docs to the setup.cfg (a remaining
question is whether the two packages should be removed from devel
group or kept in both) and makes use of this option in the
.readthedocs.yaml to make sure the theme and its dependenceies get
installed during RTD build.

The change is introduced with an expectation to remove the following
build error on ReadTheDocs:

  Could not import extension sphinxcontrib.jquery (exception: No
  module named 'sphinxcontrib.jquery')

See the build failure in eg.
https://app.readthedocs.org/projects/datalad-redcap/builds/28000704/